+++ napari/layers/labels/_tests/test_labels_mouse_bindings.py
@@ -30,7 +30,7 @@ def Event():
 )
 def test_paint(Event, brush_shape, expected_sum):
     """Test painting labels with circle/square brush."""
-    data = np.ones((20, 20))
+    data = np.ones((20, 20), dtype=np.int32)
     layer = Labels(data)
     layer.brush_size = 10
     assert layer.cursor_size == 10
@@ -67,7 +67,7 @@ def test_paint(Event, brush_shape, expected_sum):
 )
 def test_paint_scale(Event, brush_shape, expected_sum):
     """Test painting labels with circle/square brush when scaled."""
-    data = np.ones((20, 20))
+    data = np.ones((20, 20), dtype=np.int32)
     layer = Labels(data, scale=(2, 2))
     layer.brush_size = 10
     layer.brush_shape = brush_shape
@@ -103,7 +103,7 @@ def test_paint_scale(Event, brush_shape, expected_sum):
 )
 def test_erase(Event, brush_shape, expected_sum):
     """Test erasing labels with different brush shapes."""
-    data = np.ones((20, 20))
+    data = np.ones((20, 20), dtype=np.int32)
     layer = Labels(data)
     layer.brush_size = 10
     layer.mode = 'erase'
@@ -136,7 +136,7 @@ def test_erase(Event, brush_shape, expected_sum):
 
 def test_pick(Event):
     """Test picking label."""
-    data = np.ones((20, 20))
+    data = np.ones((20, 20), dtype=np.int32)
     data[:5, :5] = 2
     data[-5:, -5:] = 3
     layer = Labels(data)
@@ -160,7 +160,7 @@ def test_pick(Event):
 
 def test_fill(Event):
     """Test filling label."""
-    data = np.ones((20, 20))
+    data = np.ones((20, 20), dtype=np.int32)
     data[:5, :5] = 2
     data[-5:, -5:] = 3
     layer = Labels(data)
@@ -195,7 +195,7 @@ def test_fill(Event):
 
 def test_fill_nD_plane(Event):
     """Test filling label nD plane."""
-    data = np.ones((20, 20, 20))
+    data = np.ones((20, 20, 20), dtype=np.int32)
     data[:5, :5, :5] = 2
     data[0, 8:10, 8:10] = 2
     data[-5:, -5:, -5:] = 3
@@ -238,7 +238,7 @@ def test_fill_nD_plane(Event):
 
 def test_fill_nD_all(Event):
     """Test filling label nD."""
-    data = np.ones((20, 20, 20))
+    data = np.ones((20, 20, 20), dtype=np.int32)
     data[:5, :5, :5] = 2
     data[0, 8:10, 8:10] = 2
     data[-5:, -5:, -5:] = 3
